Create float.s
"fmult" is the most important method in this file. It is used to multiply two single precision floating point numbers in IEEE 754 format. An optimized version of this function can be fount in file floatopt.s. 

Note: This method does NOT test for overflow or underflow. Answers are not rounded.

Arguments/Parameters: 
    $a0 :    floating point number
    $a1 :    floating point number
Return:
    $v0:     contains the multiplication of $a0 and $a1 in IEEE 754 format.
Interupts:
    No interupts or tests occur.